DATAREDIS-315 - Add getClusterSlaves to ClusterCommands
We now allow reading slaves for a given master from the cluster using the CLUSTER SLAVES command. We improved parsing of the response for both CLUSTER NODES and CLUSTER SLAVES to read more information (eg. node type and masterId) as well as added support for single slot assignment of a cluster node.